+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/design/SingularFieldRule.java
@@ -4,8 +4,6 @@
 
 package net.sourceforge.pmd.lang.java.rule.design;
 
-import static net.sourceforge.pmd.lang.java.ast.JModifier.FINAL;
-import static net.sourceforge.pmd.lang.java.ast.JModifier.STATIC;
 import static net.sourceforge.pmd.util.CollectionUtil.setOf;
 
 import java.util.ArrayList;
@@ -19,9 +17,11 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import net.sourceforge.pmd.lang.java.ast.ASTAssignableExpr.ASTNamedReferenceExpr;
 import net.sourceforge.pmd.lang.java.ast.ASTBodyDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTFieldDeclaration;
+import net.sourceforge.pmd.lang.java.ast.ASTInitializer;
 import net.sourceforge.pmd.lang.java.ast.ASTLambdaExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTTypeDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTVariableId;
+import net.sourceforge.pmd.lang.java.ast.JModifier;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
 import net.sourceforge.pmd.lang.java.ast.ModifierOwner;
 import net.sourceforge.pmd.lang.java.ast.ModifierOwner.Visibility;
@@ -95,9 +95,13 @@ public class SingularFieldRule extends AbstractJavaRulechainRule {
         return null;
     }
 
-    public static boolean mayBeSingular(ModifierOwner varId) {
+    private static boolean mayBeSingular(ModifierOwner varId) {
         return varId.getEffectiveVisibility().isAtMost(Visibility.V_PRIVATE)
-            && !varId.getModifiers().hasAny(STATIC, FINAL);
+            // We ignore final variables for a reason:
+            // - if they're static they are there to share a value and that is not a mistake
+            // - if they're not static then if this rule matches, then so does FinalFieldCouldBeStatic,
+            // and that rule has the better fix.
+            && !varId.hasModifiers(JModifier.FINAL);
     }
 
     private boolean isSingularField(ASTTypeDeclaration fieldOwner, ASTVariableId varId, DataflowResult dataflow) {
@@ -105,11 +109,13 @@ public class SingularFieldRule extends AbstractJavaRulechainRule {
             return false; // don't report unused field
         }
 
+        boolean isStaticField = varId.isStatic();
         //Check usages for validity & group them by scope
         //They're valid if they don't escape the scope of their method, eg by being in a nested class or lambda
         Map<ASTBodyDeclaration, List<ASTNamedReferenceExpr>> usagesByScope = new HashMap<>();
         for (ASTNamedReferenceExpr usage : varId.getLocalUsages()) {
-            if (usage.getEnclosingType() != fieldOwner || !JavaAstUtils.isThisFieldAccess(usage)) {
+            if (usage.getEnclosingType() != fieldOwner
+                || !isStaticField && !JavaAstUtils.isThisFieldAccess(usage)) {
                 return false; // give up
             }
             ASTBodyDeclaration enclosing = getEnclosingBodyDecl(fieldOwner, usage);
@@ -121,7 +127,7 @@ public class SingularFieldRule extends AbstractJavaRulechainRule {
 
         // the field is singular if it is used as a local var in every method.
         for (ASTBodyDeclaration method : usagesByScope.keySet()) {
-            if (method != null && !usagesDontObserveValueBeforeMethodCall(usagesByScope.get(method), dataflow)) {
+            if (method != null && usagesObserveValueBeforeMethodCall(usagesByScope.get(method), dataflow)) {
                 return false;
             }
         }
@@ -129,10 +135,16 @@ public class SingularFieldRule extends AbstractJavaRulechainRule {
         return true;
     }
 
-    private @Nullable ASTBodyDeclaration getEnclosingBodyDecl(JavaNode stop, ASTNamedReferenceExpr usage) {
-        return usage.ancestors()
-                    .takeWhile(it -> it != stop)
-                    .first(ASTBodyDeclaration.class);
+    private @Nullable ASTBodyDeclaration getEnclosingBodyDecl(ASTTypeDeclaration enclosingType, ASTNamedReferenceExpr usage) {
+        ASTBodyDeclaration decl = usage.ancestors()
+                                       .takeWhile(it -> it != enclosingType)
+                                       .first(ASTBodyDeclaration.class);
+        if (decl instanceof ASTFieldDeclaration
+            || decl instanceof ASTInitializer) {
+            // then the usage is logically part of the ctors.
+            return enclosingType;
+        }
+        return decl;
     }
 
     private boolean hasEnclosingLambda(JavaNode stop, ASTNamedReferenceExpr usage) {
@@ -141,13 +153,13 @@ public class SingularFieldRule extends AbstractJavaRulechainRule {
                     .any(it -> it instanceof ASTLambdaExpression);
     }
 
-    private boolean usagesDontObserveValueBeforeMethodCall(List<ASTNamedReferenceExpr> usages, DataflowResult dataflow) {
+    private boolean usagesObserveValueBeforeMethodCall(List<ASTNamedReferenceExpr> usages, DataflowResult dataflow) {
         for (ASTNamedReferenceExpr usage : usages) {
             ReachingDefinitionSet reaching = dataflow.getReachingDefinitions(usage);
             if (reaching.containsInitialFieldValue()) {
-                return false;
+                return true;
             }
         }
-        return true;
+        return false;
     }
 }